+++ libraries/moduloMap/CHANGELOG.md
@@ -6,11 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.3] - 2025-09-09
+- add begin(float maximum) wrapper
+- add moduloMap as (derived) class
+- update GitHub actions
+- minor edits
+
 ## [0.1.2] - 2023-11-14
 - update readme.md
 - update keywords.txt
 - minor edits
-
 
 ## [0.1.1] - 2022-10-21
 - add access functions to internal variables.